requirements.txt
# Phase 1: RAG + Cache + Multi-Platform
# Install: pip install -r requirements.txt

# === Core Dependencies (existing) ===
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0

# === Web Framework (LINE Bot + Facebook) ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
line-bot-sdk>=3.9.0
pydantic>=2.0.0

# === NEW: Database (PostgreSQL) ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# === NEW: RAG Stack ===
# LlamaIndex Core
llama-index-core>=0.11.0
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openai>=0.2.0
llama-index-vector-stores-chroma>=0.2.0

# Vector Database
chromadb>=0.5.0

# Cache
redis>=5.0.0

# === Authentication (for Admin Dashboard) ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# === Optional: Enhanced Features ===
# Facebook/Instagram Scraping (existing)
apify-client>=2.4.0
playwright>=1.44.0

# Scheduling
schedule>=1.2.1

# === Development & Testing ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # For async testing

# === Production ===
gunicorn>=21.2.0
